Add files via upload
This project reflects foundational frontend development skills and an understanding of responsive design principles. While Tripzy is currently a static site, it can be extended in the future with backend functionality to support features like user authentication, dynamic booking systems, or admin dashboards. Tripzy can also serve as a base template for travel agencies or tour operators looking to establish a digital presence.

Overall, Tripzy is a demonstration of how simple frontend technologies can be combined effectively to create a polished and functional website. It is suitable for showcasing on portfolios and can be hosted using GitHub Pages or any static site hosting platform.